Add pkg/serial with a runtime poller API
The current serial libraries for Go are all written pre-1.12 and do not
make use of `syscall.SetNonblock` and the Go runtime poller.

This implementation uses the runtime poller to provide `SetReadDeadline`
and `SetWriteDeadline`, including wake-up on `Close`.

This enables us to write an Xsens client that behaves very similarly to
a network-based client, with a simple, idiomatic API.