Add Client with high-level API
The Client has come about as a result of using the lower-level API in a
few implementations.

The Client has a higher-level API for putting the Xsens in Config and
Measurement mode, and for reading measurement data.

This API makes it (arguably) easier to write downstream code, since the
higher-level API can be mocked and unit tests do not need to know about
the Xsens wire format (which is currently the case).